Organize scan logic by subject matter. This services two purposes: modularize by business concerns (scan data) and increase opportunities to run concurrent async processing. This commit includes some failing tests, apparently due to a jest import problems - these tests will be moved in a subsequent commit.
