+++ libs/core-scanner/src/core-scanner.service.ts
@@ -2,10 +2,8 @@ import { HttpService, Injectable, Logger } from '@nestjs/common';
 
 import { BrowserService } from '@app/browser';
 import { CoreInputDto } from '@app/core-scanner/core.input.dto';
-import {
-  buildErrorResult,
-  coreScan,
-} from '@app/page-scanners/home-page/core-scan';
+import { coreScan } from '@app/page-scanners/home-page/core-scan';
+import { buildCoreErrorResult } from '@app/page-scanners/scans/core';
 import { Scanner } from 'libs/scanner.interface';
 import { CoreResult } from 'entities/core-result.entity';
 import { ScanStatus } from './scan-status';
@@ -21,16 +19,18 @@ export class CoreScannerService implements Scanner<CoreInputDto, CoreResult> {
 
   async scan(input: CoreInputDto): Promise<CoreResult> {
     try {
-      return await this.browserService.processPage<CoreResult>((page) => {
-        const result = coreScan(this.httpService, this.logger, input, page);
-        this.logger.log({ msg: 'core scan results', ...input });
-        return result;
-      });
+      const result = await this.browserService.processPage<CoreResult>(
+        (page) => {
+          return coreScan(this.httpService, this.logger, input, page);
+        },
+      );
+      this.logger.log({ msg: 'core scan results', ...input });
+      return result;
     } catch (error) {
       const err = error as Error;
 
       // build error result on error
-      const result = buildErrorResult(input, err);
+      const result = buildCoreErrorResult(input, err);
 
       // log if the error is unknown
       if (result.status == ScanStatus.UnknownError) {
