+++ libs/database/src/core-results/core-result.service.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { CoreResult } from 'entities/core-result.entity';
 import { Website } from 'entities/website.entity';
-import { mock, MockProxy, mockReset } from 'jest-mock-extended';
+import { mock, mockReset } from 'jest-mock-extended';
 import { Repository } from 'typeorm';
 import { CoreResultService } from './core-result.service';
