+++ libs/page-scanners/src/scans/core.test.ts
@@ -1,0 +1,3 @@
+describe('core scan', () => {
+  xit('works', () => {});
+});
