+++ libs/page-scanners/src/scans/not-found.test.ts
@@ -1,0 +1,3 @@
+describe('not found scan', () => {
+  xit('works', () => {});
+});
