+++ libs/page-scanners/src/scans/not-found.ts
@@ -1,0 +1,26 @@
+import { HttpService, HttpStatus } from '@nestjs/common';
+import { Agent } from 'https';
+import { v4 } from 'uuid';
+
+export const notFoundTest = async (
+  httpService: HttpService,
+  url: string,
+): Promise<boolean> => {
+  const randomUrl = new URL(url);
+  randomUrl.pathname = `not-found-test${v4()}`;
+
+  const agent = new Agent({
+    rejectUnauthorized: false,
+  });
+
+  const resp = await httpService
+    .get(randomUrl.toString(), {
+      validateStatus: (_) => {
+        return true;
+      },
+      httpsAgent: agent,
+    })
+    .toPromise();
+
+  return resp.status == HttpStatus.NOT_FOUND;
+};
