+++ libs/page-scanners/src/scans/seo.test.ts
@@ -1,0 +1,3 @@
+describe('seo scan', () => {
+  xit('works', () => {});
+});
