+++ libs/page-scanners/src/scans/third-party.test.ts
@@ -1,0 +1,3 @@
+describe('third-party scan', () => {
+  xit('works', () => {});
+});
