+++ libs/page-scanners/src/scans/uswds.test.ts
@@ -1,0 +1,3 @@
+describe('uswds scan', () => {
+  xit('works', () => {});
+});
